Fix PROGMEM errors and warnings

+++ tlc_config.h
@@ -158,7 +158,9 @@
 /** Arranges 2 grayscale values (0 - 4095) in the packed array format (3 bytes).
     This is for array initializers only: the output is three comma seperated
     8-bit values. */
-#define GS_DUO(a, b)    ((a) >> 4), ((a) << 4) | ((b) >> 8), (b)
+//#define GS_DUO(a, b)  ((a) >> 4), ((a) << 4) | ((b) >> 8), (b)
+#define GS_DUO(a, b)    (uint8_t)((a) >> 4), (uint8_t)((a) << 4) | (uint8_t)((b) >> 8), (uint8_t)(b)
+
 
 
 #if VPRG_ENABLED
